removed feed file and xfer code to landing page component - feed.js, my bad guys
